Remove the use of -S in fcopy invocations
This branch is expected to remain compatible with the FAI package versions
available in stretch, so we can't use that new feature. Presently none of the
installed files rely on fcopy's variable substitution functionality, so the
change was simply a matter of removing the option from the command line
invocation.