Add Octavia Amphora image building
Octavia is OpenStack Load Balancer as a Service. It needs an image
running octavia-agent which is a REST API for configuring haproxy
inside the image. All is already packaged in Bullseye/Sid, so it
should just work there, but the Buster package should be updated
(the SRM must allow an update of Octavia before this image can be
built for Buster).